+++ react-components/.storybook/useDayjsLocale.js
@@ -1,0 +1,3 @@
+import '../src/utils/dayjsLocale';
+
+export const useDayjsLocale = (StoryFn) => StoryFn();