+++ react-components/src/components/Table/model/i-cell-instance.ts
@@ -1,4 +1,5 @@
-import { TableColumn } from 'tetacom/react-components';
+import { IDictionary, TableColumn } from 'tetacom/react-components';
+import { Row } from '@tanstack/react-table';
 
 export interface ICellInstance<T> {
   row: T;
@@ -13,6 +14,8 @@ export interface ICellInstanceValue<T> extends ICellInstance<T> {
   value: any;
 }
 
-export interface ICustomCell {
+export interface ICustomCell<T> {
   value: any;
+  row: Row<T>;
+  dict?: IDictionary | null;
 }