+++ react-components/src/components/Tooltip/model/index.ts
@@ -16,4 +16,10 @@ export interface TooltipProps extends PropsWithChildren {
 
   // Событие для показа подсказки
   target?: 'click' | 'hover';
+
+  // Задержка отображения подсказки в миллисекундах
+  delay?: number;
+
+  // Показывать всплывающее окно рядом с указателм мыши
+  mouseFollow?: boolean;
 }